requirements.txt
requests==2.31.0
beautifulsoup4==4.12.0
PyMuPDF==1.24.0
python-dotenv==1.0.0
google-cloud-aiplatform==1.70.0
langchain==0.3.0
langchain-google-vertexai==2.0.0
tiktoken==0.7.0